+++ NEWS.md
@@ -5,6 +5,9 @@
   `@inherit my_fun` will inherit all, or you can select specific tags to 
   inherit with `@inherit my_fun return params` (#384).
 
+* New `@inheritSection fun title` allows you to inherit the contents of 
+  a single section from another topic (#513).
+
 * `@inheritParams` now works recursively, so that you can inherit parameters
   from a function that inherited its paramters from somewhere else.